code refactoring

+++ server/routes/labels.js
@@ -1,4 +1,4 @@
-import i18next from "i18next";
+import i18next from 'i18next';
 
 export default (app) => {
   app
@@ -80,17 +80,28 @@ export default (app) => {
     .delete('/labels/:id', { name: 'deletelabel' }, async (req, reply) => {
       const { id } = req.params;
       try {
-        const label = await app.objection.models.label.query().findById(id);
+        const label = await app.objection.models.label
+          .query()
+          .findById(id)
+          .withGraphFetched('tasks');
+
         if (!label) {
           req.flash('error', i18next.t('flash.labels.delete.notFound'));
-          return reply.label(404).send('label not found');
+          return reply.status(404).send('Label not found');
         }
+
+        if (label.tasks.length > 0) {
+          // deny delete of label if it is connected to a task
+          req.flash('error', i18next.t('flash.labels.delete.hasTasks'));
+          return reply.redirect('/labels');
+        }
+
         await label.$query().delete();
         req.flash('info', i18next.t('flash.labels.delete.success'));
-        reply.redirect('/labels');
+        return reply.redirect('/labels');
       } catch (error) {
         req.flash('error', i18next.t('flash.labels.delete.error'));
-        return reply.label(500).send('Internal Server Error');
+        return reply.status(500).send('Internal Server Error');
       }
     });
 };
